add reversal weight

+++ app/services/crypto/analysis/ignored_pattern.go
@@ -10,5 +10,5 @@ func isInAboveUpperBandAndDownTrend(result *models.BandResult) bool {
 	lastBand := result.Bands[len(result.Bands)-1]
 	lastFourData := result.Bands[len(result.Bands)-4 : len(result.Bands)]
 
-	return int8(lastBand.Candle.Low) > int8(lastBand.Upper) && CalculateTrends(lastFourData) == models.TREND_DOWN
+	return float64(lastBand.Candle.Low) > lastBand.Upper && CalculateTrends(lastFourData) == models.TREND_DOWN
 }
